Add files via upload
The Loan Eligibility Prediction System is designed to predict the loan status of students by utilizing various inputs. This project is developed using the machine learning algorithm logistic regression. The primary objective is to automate the loan eligibility process within the banking sector. By employing logistic regression, the system evaluates the provided inputs to determine whether a student is eligible for a loan or not. This automation streamlines and enhances the efficiency of the loan eligibility assessment process in the banking sector